Update wallet.js
Key Change:
Mobile detection: We added a state isMobile to detect whether the device is mobile. We check for the "Mobi" or "Android" keyword in the navigator.userAgent to set this.
Mobile-specific behavior: If the device is detected as mobile (isMobile is true), we show an alert message informing the user that the wallet connection is not supported on mobile for now.